Remove old leaflet examples
